+++ packages/swapper/src/swappers/cow/CowApproveInfinite/CowApproveInfinite.ts
@@ -1,5 +1,3 @@
-import { SupportedChainIds } from '@shapeshiftoss/types'
-
 import { ApproveInfiniteInput, SwapError, SwapErrorTypes } from '../../../api'
 import { erc20Abi } from '../../utils/abi/erc20-abi'
 import { grantAllowance } from '../../utils/helpers/helpers'
@@ -7,8 +5,8 @@ import { CowSwapperDeps } from '../CowSwapper'
 import { MAX_ALLOWANCE } from '../utils/constants'
 
 export async function CowApproveInfinite(
-  { adapterManager, web3 }: CowSwapperDeps,
-  { quote, wallet }: ApproveInfiniteInput<SupportedChainIds>
+  { adapter, web3 }: CowSwapperDeps,
+  { quote, wallet }: ApproveInfiniteInput<'eip155:1'>
 ) {
   try {
     const allowanceGrantRequired = await grantAllowance({
@@ -17,7 +15,7 @@ export async function CowApproveInfinite(
         sellAmount: MAX_ALLOWANCE
       },
       wallet,
-      adapterManager,
+      adapter,
       erc20Abi,
       web3
     })