+++ packages/swapper/src/swappers/cow/utils/helpers/helpers.test.ts
@@ -1,4 +1,4 @@
-import { ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
+import { ethereum } from '@shapeshiftoss/chain-adapters'
 import Web3 from 'web3'
 
 import { ETH, FOX, WBTC } from '../../../utils/test-data/assets'
@@ -11,7 +11,7 @@ jest.mock('../cowService')
 describe('utils', () => {
   const cowSwapperDeps: CowSwapperDeps = {
     apiUrl: 'https://api.cow.fi/mainnet/api/',
-    adapterManager: <ChainAdapterManager>{},
+    adapter: <ethereum.ChainAdapter>{},
     web3: <Web3>{}
   }