+++ packages/swapper/src/swappers/utils/helpers/helpers.test.ts
@@ -25,12 +25,12 @@ Web3.mockImplementation(() => ({
 
 describe('utils', () => {
   const { tradeQuote, sellAsset } = setupQuote()
-  const { web3Instance, adapterManager } = setupDeps()
+  const { web3, adapter } = setupDeps()
 
   describe('getAllowanceRequired', () => {
     const getAllowanceInput = {
       receiveAddress: '0x0',
-      web3: web3Instance,
+      web3,
       erc20AllowanceAbi,
       allowanceContract: '0x0',
       sellAmount: '100',
@@ -48,7 +48,7 @@ describe('utils', () => {
 
     it('should return sellAmount if allowanceOnChain is 0', async () => {
       const allowanceOnChain = '0'
-      ;(web3Instance.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
+      ;(web3.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
         methods: {
           allowance: jest.fn(() => ({
             call: jest.fn(() => allowanceOnChain)
@@ -63,7 +63,7 @@ describe('utils', () => {
 
     it('should throw error if allowanceOnChain is undefined', async () => {
       const allowanceOnChain = undefined
-      ;(web3Instance.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
+      ;(web3.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
         methods: {
           allowance: jest.fn(() => ({
             call: jest.fn(() => allowanceOnChain)
@@ -79,7 +79,7 @@ describe('utils', () => {
     it('should return 0 if sellAmount minus allowanceOnChain is negative', async () => {
       const allowanceOnChain = '1000'
 
-      ;(web3Instance.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
+      ;(web3.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
         methods: {
           allowance: jest.fn(() => ({
             call: jest.fn(() => allowanceOnChain)
@@ -93,7 +93,7 @@ describe('utils', () => {
     it('should return sellAsset minus allowanceOnChain', async () => {
       const allowanceOnChain = '100'
 
-      ;(web3Instance.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
+      ;(web3.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
         methods: {
           allowance: jest.fn(() => ({
             call: jest.fn(() => allowanceOnChain)
@@ -118,7 +118,7 @@ describe('utils', () => {
       const quote = {
         ...tradeQuote
       }
-      ;(web3Instance.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
+      ;(web3.eth.Contract as jest.Mock<unknown>).mockImplementation(() => ({
         methods: {
           approve: jest.fn(() => ({
             encodeABI: jest.fn(
@@ -128,8 +128,10 @@ describe('utils', () => {
         }
       }))
 
+      ;(adapter.buildSendTransaction as jest.Mock).mockResolvedValueOnce({ txToSign: {} })
+      ;(adapter.broadcastTransaction as jest.Mock).mockResolvedValueOnce('broadcastedTx')
       expect(
-        await grantAllowance({ quote, wallet, adapterManager, erc20Abi, web3: web3Instance })
+        await grantAllowance({ quote, wallet, adapter, erc20Abi, web3 })
       ).toEqual('broadcastedTx')
     })
   })